refactor, style: 중복 코드 dto 및 메서드로 분리, 불필요한 import 삭제 및 코드 정렬

+++ src/main/java/com/deal4u/fourplease/domain/auction/dto/BidSummaryDto.java
@@ -7,10 +7,10 @@ public record BidSummaryDto(
         BigDecimal maxPrice,
         int bidCount
 ) {
-    public static BidSummaryDto toBidSummaryDto(List<Long> bidList) {
+    public static BidSummaryDto toBidSummaryDto(List<BigDecimal> bidList) {
         if (bidList.isEmpty()) {
             return new BidSummaryDto(BigDecimal.ZERO, 0);
         }
-        return new BidSummaryDto(BigDecimal.valueOf(bidList.getFirst()), bidList.size());
+        return new BidSummaryDto(bidList.getFirst(), bidList.size());
     }
 }
